Keep span when reordering grid items (#6642)
**Problem:**

Spans (or otherwise 1+ sized elements) are not preserved when using
`gridReorderStrategy`.

**Fix:**

When running the reorder strategy, instead of nuking all positioning
props only nuke the ones which are not used and, for elements that
occupy multiple cells, set the respective width or height using the
`span` syntax on the correct shorthand.

I also took this chance to move the reorder tests to their own file (and
added the new tests there).

Fixes #6641
